+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## Release 1.1.0
+
+### New features
+- New block "CollectData" to collect data multiple times to forward them in a single event
+- New block "Log" to log received content (usable to debug)
+
+### Improvements
+- Increased number of JSON key/sources from 4 to 6
+
 ## Release 1.0.0
 - Initial commit